# Optimize event matching in DiagnosticsObserver